url-service/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
mysql-connector-python==8.0.26
python-dotenv==0.19.0
gunicorn==20.1.0
werkzeug==2.0.1 